# docs/swagger.yaml
definitions:
...
  controllers.specInput:
    properties:
      camera:
        type: string
      memory:
        type: string
      phone_id:
        type: integer
      price:
        type: string
      processor:
        type: string
      review:
        type: string
      screen:
        type: string
      storage:
        type: string
    type: object
...
  models.Spec:
    properties:
      camera:
        type: string
      created_at:
        type: string
      id:
        type: integer
      memory:
        type: string
      phone_id:
        type: integer
      price:
        type: string
      processor:
        type: string
      review:
        type: string
      screen:
        type: string
      storage:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /specs:
    get:
      description: Get a list of Spec.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Spec'
            type: array
      summary: Get all Spec.
      tags:
      - Spec
    post:
      description: Creating a new Spec.
      parameters:
      - description: the body to create a new Spec
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.specInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Spec'
      security:
      - BearerToken: []
      summary: Create New Spec. (Admin only)
      tags:
      - Spec
  /specs/{id}:
    delete:
      description: Delete a Spec by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Spec id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Spec. (Admin only)
      tags:
      - Spec
    get:
      description: Get an Spec by id.
      parameters:
      - description: Spec id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Spec'
      summary: Get Spec.
      tags:
      - Spec
    patch:
      description: Update Spec by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Spec id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update age spec category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.specInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Spec'
      security:
      - BearerToken: []
      summary: Update Spec. (Admin only)
      tags:
      - Spec
swagger: "2.0"